+++ modules/compiler/utils/source/define_mux_dma_pass.cpp
@@ -16,17 +16,6 @@
 
 #include <compiler/utils/builtin_info.h>
 #include <compiler/utils/define_mux_dma_pass.h>
-#include <compiler/utils/dma.h>
-#include <compiler/utils/pass_functions.h>
-#include <llvm/IR/Argument.h>
-#include <llvm/IR/CallingConv.h>
-#include <llvm/IR/Constants.h>
-#include <llvm/IR/IRBuilder.h>
-#include <llvm/IR/Instructions.h>
-#include <llvm/IR/Module.h>
-#include <llvm/Pass.h>
-#include <llvm/Support/raw_ostream.h>
-#include <multi_llvm/opaque_pointers.h>
 
 #define DEBUG_TYPE "define-mux-dma"
 
@@ -37,7 +26,7 @@ PreservedAnalyses compiler::utils::DefineMuxDmaPass::run(
   bool Changed = false;
   auto &BI = AM.getResult<BuiltinInfoAnalysis>(M);
 
-  // Define all mux builtins
+  // Define all mux dma builtins
   for (auto &F : M.functions()) {
     auto ID = BI.analyzeBuiltin(F).ID;
     if (!BI.isMuxDmaBuiltinID(ID)) {